Fix editor state and operation handling in Laboratory + e2e tests (#6282)
Fixes

When opening a new tab or selecting one of the saved operations a wrong query was populated. It was always the default query or the one that's active. Meaning, you couldn't select and see the saved operation :)

When saving the operation, the submit button of the form was always disabled, even when the state of the form was valid.

e2e tests

Added tests for CRUD of collections and their operations.
The scenario where a user visits a shared link to an operation is also now tested.