Assign custom roles to members without a role (#6332)
This migration is going to create a new role for each group of members that have the same scopes but no role assigned.

Users won't be affected by this change, as they will still have the same scopes.

The role will be named Auto Role {counter}.
The counter will be reset for each organization.

The script will create 1 or 2 roles for 97% of organizations that haven't complete the migration.

Completes https://the-guild.dev/graphql/hive/product-updates/2023-12-05-member-roles